EDCimport Database

This class of object represents a database, as the result of an EDCimport reading function. It has its own print() method.

Functions returning edc_database objects

As per now, reading functions are: read_trialmaster(), read_all_sas(), read_all_xpt(), and read_all_csv().

Structure

While it is not usually usefull to query them, an edc_database object is a named list containing:

all the datasets from the source files

datetime_extraction and date_extraction the inferred date of data extraction

.lookup a temporary copy of the lookup table

See also