Extract an archive using 7zip

Use 7zip to extract an archive. Obviously, 7zip should be installed beforehand.

extract_7z(archive, target_dir, password = NULL, path_7zip = NULL)

Arguments

archive

the archive file

target_dir

the target directory

password

the password of the archive, if any

path_7zip

See section below. Default to get_7z_dir().

Value

the success/error message. Mainly used for its side effect of extracting the archive.

Install 7-zip

If 7-zip is not installed on your computer, the easiest way to get it is to run installr::install.7zip().

Add 7zip to the PATH

For this function to work, 7-zip should be registered as a PATH environment variable. There are many resources online to help you register if, e.g. https://www.java.com/en/download/help/path.html. If you cannot change the PATH on your computer, then you can add the path to 7zip executable directory in the parameter path_7zip or globally using options(path_7zip="path/to/7zip"). This will add 7zip to the PATH for the current session.

See also