Get a table with the latest date for each patient

This function search for date columns in every tables and returns the latest date for each patient with the variable it comes from. Useful in survival analysis to get the right censoring time.

Usage

lastnews_table(
  except = NULL,
  with_ties = FALSE,
  show_delta = FALSE,
  numeric_id = TRUE,
  prefer = NULL,
  regex = FALSE,
  warn_if_future = TRUE
)

Arguments

except

the datasets/columns that should not be searched. Example: a scheduled visit for which the patient may have died before attending should not be considered.

with_ties

in case of tie, whether to return the first origin (FALSE) or all the origins that share this tie (TRUE).

show_delta

whether to compute the difference between the last prefer date and the actual last date

numeric_id

set to FALSE if the patient ID column is not numeric

prefer

preferred origins in the event of a tie. Usually the followup table.

regex

whether to consider except and prefer as regex.

warn_if_future

whether to show a warning about dates that are after the extraction date. Can also be a csv file path to save the warning as csv (see csv_path argument in edc_data_warn).

Value

a dataframe

Examples

db = edc_example()
#> Warning: Option "edc_lookup" has been overwritten.
load_database(db)
lastnews_table()
#> # A tibble: 50 × 5
#>    subjid last_date           origin_data origin_col origin_label    
#>     <dbl> <dttm>              <chr>       <chr>      <chr>           
#>  1      1 2010-08-01 18:59:37 data2       date4      Date at visit 4 
#>  2      2 2010-07-31 15:32:45 data2       date4      Date at visit 4 
#>  3      3 2010-07-22 11:24:37 data2       date5      Date at visit 5 
#>  4      4 2010-07-23 20:38:32 data3       date10     Date at visit 10
#>  5      5 2010-07-15 07:09:47 data3       date10     Date at visit 10
#>  6      6 2010-07-20 12:27:00 data3       date10     Date at visit 10
#>  7      7 2010-07-28 16:24:09 data3       date9      Date at visit 9 
#>  8      8 2010-07-19 15:24:18 data3       date9      Date at visit 9 
#>  9      9 2010-08-11 03:48:27 data3       date9      Date at visit 9 
#> 10     10 2010-07-30 20:41:23 data3       date10     Date at visit 10
#> # ℹ 40 more rows
lastnews_table(except="db3")
#> # A tibble: 50 × 5
#>    subjid last_date           origin_data origin_col origin_label    
#>     <dbl> <dttm>              <chr>       <chr>      <chr>           
#>  1      1 2010-08-01 18:59:37 data2       date4      Date at visit 4 
#>  2      2 2010-07-31 15:32:45 data2       date4      Date at visit 4 
#>  3      3 2010-07-22 11:24:37 data2       date5      Date at visit 5 
#>  4      4 2010-07-23 20:38:32 data3       date10     Date at visit 10
#>  5      5 2010-07-15 07:09:47 data3       date10     Date at visit 10
#>  6      6 2010-07-20 12:27:00 data3       date10     Date at visit 10
#>  7      7 2010-07-28 16:24:09 data3       date9      Date at visit 9 
#>  8      8 2010-07-19 15:24:18 data3       date9      Date at visit 9 
#>  9      9 2010-08-11 03:48:27 data3       date9      Date at visit 9 
#> 10     10 2010-07-30 20:41:23 data3       date10     Date at visit 10
#> # ℹ 40 more rows
lastnews_table(except="db3$date9")
#> # A tibble: 50 × 5
#>    subjid last_date           origin_data origin_col origin_label    
#>     <dbl> <dttm>              <chr>       <chr>      <chr>           
#>  1      1 2010-08-01 18:59:37 data2       date4      Date at visit 4 
#>  2      2 2010-07-31 15:32:45 data2       date4      Date at visit 4 
#>  3      3 2010-07-22 11:24:37 data2       date5      Date at visit 5 
#>  4      4 2010-07-23 20:38:32 data3       date10     Date at visit 10
#>  5      5 2010-07-15 07:09:47 data3       date10     Date at visit 10
#>  6      6 2010-07-20 12:27:00 data3       date10     Date at visit 10
#>  7      7 2010-07-28 16:24:09 data3       date9      Date at visit 9 
#>  8      8 2010-07-19 15:24:18 data3       date9      Date at visit 9 
#>  9      9 2010-08-11 03:48:27 data3       date9      Date at visit 9 
#> 10     10 2010-07-30 20:41:23 data3       date10     Date at visit 10
#> # ℹ 40 more rows
lastnews_table(prefer="date10", show_delta=TRUE) 
#> # A tibble: 50 × 8
#>    subjid last_date           origin_data origin_col origin_label    
#>     <dbl> <dttm>              <chr>       <chr>      <chr>           
#>  1      1 2010-08-01 18:59:37 data3       date10     Date at visit 10
#>  2      2 2010-07-31 15:32:45 data3       date10     Date at visit 10
#>  3      3 2010-07-22 11:24:37 data3       date10     Date at visit 10
#>  4      4 2010-07-23 20:38:32 data3       date10     Date at visit 10
#>  5      5 2010-07-15 07:09:47 data3       date10     Date at visit 10
#>  6      6 2010-07-20 12:27:00 data3       date10     Date at visit 10
#>  7      7 2010-07-28 16:24:09 data3       date9      Date at visit 9 
#>  8      8 2010-07-19 15:24:18 data3       date9      Date at visit 9 
#>  9      9 2010-08-11 03:48:27 data3       date9      Date at visit 9 
#> 10     10 2010-07-30 20:41:23 data3       date10     Date at visit 10
#> # ℹ 40 more rows
#> # ℹ 3 more variables: preferred_last_date <dttm>, preferred_origin <chr>,
#> #   delta <drtn>
lastnews_table() %>% 
  dplyr::count(origin = glue::glue("{origin_data}${origin_col}"), 
  sort=TRUE)
#> # A tibble: 5 × 2
#>   origin           n
#>   <glue>       <int>
#> 1 data3$date10    36
#> 2 data3$date9     10
#> 3 data2$date4      2
#> 4 data2$date5      1
#> 5 data3$date8      1

csv_file = tempfile(fileext=".csv")
lastnews_table(prefer="date9", warn_if_future=csv_file) 
#> # A tibble: 50 × 5
#>    subjid last_date           origin_data origin_col origin_label    
#>     <dbl> <dttm>              <chr>       <chr>      <chr>           
#>  1      1 2010-08-01 18:59:37 data2       date4      Date at visit 4 
#>  2      2 2010-07-31 15:32:45 data2       date4      Date at visit 4 
#>  3      3 2010-07-22 11:24:37 data2       date5      Date at visit 5 
#>  4      4 2010-07-23 20:38:32 data3       date10     Date at visit 10
#>  5      5 2010-07-15 07:09:47 data3       date10     Date at visit 10
#>  6      6 2010-07-20 12:27:00 data3       date10     Date at visit 10
#>  7      7 2010-07-28 16:24:09 data3       date9      Date at visit 9 
#>  8      8 2010-07-19 15:24:18 data3       date9      Date at visit 9 
#>  9      9 2010-08-11 03:48:27 data3       date9      Date at visit 9 
#> 10     10 2010-07-30 20:41:23 data3       date10     Date at visit 10
#> # ℹ 40 more rows