Split mixed datasets

Split mixed tables, i.e. tables that hold both long data (N values per patient) and short data (one value per patient, duplicated on N lines), into one long table and one short table.

split_mixed_datasets(
  datasets = get_datasets(),
  id,
  ...,
  verbose = TRUE,
  output_code = FALSE
)

Arguments

datasets

the datasets to consider. Use the helper get_datasets() if needed.

id

the patient identifier, probably "SUBJID". Should be shared by all datasets.

...

not used

verbose

whether to print informations about the process.

output_code

whether to print the code to explictly write. Can also be a file path.

Value

a list of the new long and short tables. Use load_list() to load them in the global environment.

Examples

#tm = read_trialmaster("filename.zip", pw="xx")
tm = edc_example_mixed()
names(tm)
#> [1] "short"               "long_pure"           "long_mixed"         
#> [4] "date_extraction"     "datetime_extraction" ".lookup"            
load_list(tm)
print(long_mixed) #`val1` and `val2` are long but `val3` is short
#> # A tibble: 200 × 4
#>    SUBJID   val1  val2 val3 
#>     <int>  <dbl> <dbl> <chr>
#>  1      1 -1.01   9.08 B    
#>  2      1  1.31   9.45 B    
#>  3      2  3.40  11.7  C    
#>  4      2 -1.38   9.28 C    
#>  5      3  1.53   9.39 D    
#>  6      3 -0.398  8.85 D    
#>  7      4  1.13  10.9  E    
#>  8      4  0.204 10.1  E    
#>  9      5  0.244 11.3  F    
#> 10      5  0.976 10.6  F    
#> # ℹ 190 more rows

mixed_data = split_mixed_datasets(tm, id="SUBJID")
#> Error in eval(expr, envir, enclos): object 'tm' not found
load_list(mixed_data)
#> Error in eval(expr, envir, enclos): object 'mixed_data' not found
print(long_mixed_short) 
#> Error in eval(expr, envir, enclos): object 'long_mixed_short' not found
print(long_mixed_long) 
#> Error in eval(expr, envir, enclos): object 'long_mixed_long' not found

#alternatively, get the code and only use the datasets you need
split_mixed_datasets(tm, id="SUBJID", output_code=TRUE)
#> Error in eval(expr, envir, enclos): object 'tm' not found
split_mixed_datasets(tm, id="SUBJID", output_code="mixed_code.R")
#> Error in eval(expr, envir, enclos): object 'tm' not found