Split mixed datasets

Split mixed tables, i.e. tables that hold both long data (N values per patient) and short data (one value per patient, duplicated on N lines), into one long table and one short table.

split_mixed_datasets(
  datasets = get_datasets(),
  id,
  ...,
  ignore_cols = getOption("edc_crfname", "CRFNAME"),
  output_code = FALSE,
  verbose = TRUE
)

Arguments

datasets

the datasets to consider. Use the helper get_datasets() if needed.

id

the patient identifier, probably "SUBJID". Should be shared by all datasets.

...

not used

ignore_cols

columns to ignore when considering a table as long. default to

output_code

whether to print the code to explicitly write. Can also be a file path.

verbose

whether to print informations about the process.

Value

a list of the new long and short tables. Use load_list() to load them into the global environment.

Examples

#tm = read_trialmaster("filename.zip", pw="xx")
tm = edc_example_mixed()
names(tm)
#> [1] "short"               "long_pure"           "long_mixed"         
#> [4] "date_extraction"     "datetime_extraction" ".lookup"            
#load_list(tm)
print(tm$long_mixed) #`val1` and `val2` are long but `val3` is short
#> # A tibble: 200 × 5
#>    SUBJID crfname                    val1  val2 val3 
#>     <int> <chr>                     <dbl> <dbl> <chr>
#>  1      1 both short and long data -0.242  9.15 B    
#>  2      1 both short and long data  0.572 10.1  B    
#>  3      2 both short and long data  0.706  8.67 C    
#>  4      2 both short and long data  0.205 12.4  C    
#>  5      3 both short and long data  0.552  7.50 D    
#>  6      3 both short and long data -0.569 12.2  D    
#>  7      4 both short and long data  0.193 10.6  E    
#>  8      4 both short and long data  0.621  9.42 E    
#>  9      5 both short and long data -0.107 10.3  F    
#> 10      5 both short and long data -0.124  9.95 F    
#> # ℹ 190 more rows

mixed_data = split_mixed_datasets(tm, id="SUBJID", verbose=TRUE)
#> ! "SUBJID" was not found in 1 table:
#>   ".lookup".
#> ✔ There was 1 short table:
#>   "short"
#> ✔ There were 2 mixed (short+long) tables:
#>   "long_pure" and "long_mixed"
#> → Use `EDCimport::load_list()` on the result to get separated long and short
#>   data.
load_list(mixed_data)
print(long_mixed_short) 
#> # A tibble: 100 × 3
#>    SUBJID crfname                  val3 
#>     <int> <chr>                    <chr>
#>  1      1 both short and long data B    
#>  2      2 both short and long data C    
#>  3      3 both short and long data D    
#>  4      4 both short and long data E    
#>  5      5 both short and long data F    
#>  6      6 both short and long data G    
#>  7      7 both short and long data H    
#>  8      8 both short and long data I    
#>  9      9 both short and long data J    
#> 10     10 both short and long data K    
#> # ℹ 90 more rows
print(long_mixed_long) 
#> # A tibble: 200 × 3
#>    SUBJID   val1  val2
#>     <int>  <dbl> <dbl>
#>  1      1 -0.242  9.15
#>  2      1  0.572 10.1 
#>  3      2  0.706  8.67
#>  4      2  0.205 12.4 
#>  5      3  0.552  7.50
#>  6      3 -0.569 12.2 
#>  7      4  0.193 10.6 
#>  8      4  0.621  9.42
#>  9      5 -0.107 10.3 
#> 10      5 -0.124  9.95
#> # ℹ 190 more rows

#alternatively, get the code and only use the datasets you need
split_mixed_datasets(tm, id="SUBJID", output_code=TRUE)
#> ! "SUBJID" was not found in 1 table:
#>   ".lookup".
#> ✔ There was 1 short table:
#>   "short"
#> ✔ There were 2 mixed (short+long) tables:
#>   "long_pure" and "long_mixed"
#> → Copy the following code in your script to separate long and short data:
#> ## `long_pure` (dim=300x4) ---- 
#> 
#> long_pure_short = long_pure %>% 
#>   select(SUBJID, crfname) %>% 
#>   group_by(SUBJID) %>% 
#>   summarise(across(everything(), unify)) #dim=100x2 
#> 
#>  long_pure_long = long_pure %>% 
#>   select(SUBJID, val1, val2) #dim=300x3
#> 
#> 
#> ## `long_mixed` (dim=200x5) ---- 
#> 
#> long_mixed_short = long_mixed %>% 
#>   select(SUBJID, crfname, val3) %>% 
#>   group_by(SUBJID) %>% 
#>   summarise(across(everything(), unify)) #dim=100x3 
#> 
#>  long_mixed_long = long_mixed %>% 
#>   select(SUBJID, val1, val2) #dim=200x3 
filename = tempfile("mixed_code", fileext=".R")
split_mixed_datasets(tm, id="SUBJID", output_code=filename)
#> ! "SUBJID" was not found in 1 table:
#>   ".lookup".
#> ✔ There was 1 short table:
#>   "short"
#> ✔ There were 2 mixed (short+long) tables:
#>   "long_pure" and "long_mixed"
#> → Copy the code from /tmp/Rtmpo3G4Z4/mixed_code16762e5a6a2e.R in your script to
#>   separate long and short data:
#>   `utils::browseURL(/tmp/Rtmpo3G4Z4/mixed_code16762e5a6a2e.R)`
readLines(filename)
#>  [1] "## `long_pure` (dim=300x4) ---- "                    
#>  [2] ""                                                    
#>  [3] "long_pure_short = long_pure %>% "                    
#>  [4] "  select(SUBJID, crfname) %>% "                      
#>  [5] "  group_by(SUBJID) %>% "                             
#>  [6] "  summarise(across(everything(), unify)) #dim=100x2 "
#>  [7] ""                                                    
#>  [8] " long_pure_long = long_pure %>% "                    
#>  [9] "  select(SUBJID, val1, val2) #dim=300x3"             
#> [10] ""                                                    
#> [11] ""                                                    
#> [12] "## `long_mixed` (dim=200x5) ---- "                   
#> [13] ""                                                    
#> [14] "long_mixed_short = long_mixed %>% "                  
#> [15] "  select(SUBJID, crfname, val3) %>% "                
#> [16] "  group_by(SUBJID) %>% "                             
#> [17] "  summarise(across(everything(), unify)) #dim=100x3 "
#> [18] ""                                                    
#> [19] " long_mixed_long = long_mixed %>% "                  
#> [20] "  select(SUBJID, val1, val2) #dim=200x3 "